💄 Split login widgets into seperate classes and files
