+++ libs/authentication/pubspec.yaml
@@ -38,8 +38,8 @@ dev_dependencies:
 flutter:
 
   # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
+  assets:
+    - ../assets/logo.png
   #   - images/a_dot_ham.jpeg
   #
   # For details regarding assets in packages, see
